immutability: make ClusterResourceSetBinding spec.clusterName immutable via CEL
Replace the webhook-based clusterName immutability check on
ClusterResourceSetBinding with a CEL XValidation transition rule, so
it is enforced by the API server even when the validating webhook is
not running.

The rule allows the first transition away from the empty string, same
as the existing webhook behavior. This was the only validation rule
on this resource, so the webhook now only exists to satisfy the
admission.Validator interface and its dedicated test file is removed.
