+++ api/addons/v1beta1/clusterresourcesetbinding_types.go
@@ -198,6 +198,7 @@ type ClusterResourceSetBindingSpec struct {
 	// +optional
 	// +kubebuilder:validation:MinLength=1
 	// +kubebuilder:validation:MaxLength=63
+	// +kubebuilder:validation:XValidation:rule="oldSelf == '' || oldSelf == self",message="field is immutable"
 	ClusterName string `json:"clusterName,omitempty"`
 }
